fix: Long으로 타입변경

+++ backend/src/main/java/com/killseat/payment/dto/PortOneCancelResult.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class PortOneCancelResult {
-    private Integer cancelAmount;
+    private Long cancelAmount;
     private String message;
 }
